backend/agent-voice/pyproject.toml
[project]
name = "agent-voice"
version = "0.1.0"
description = "Voice intake — transcribe user monologue and extract profile."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.50.0",
  "httpx>=0.27.2",
  "python-dotenv>=1.0.1",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.0.0",
]

[project.scripts]
agent-voice = "agent_voice.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true
